Create registration composable
